Fix ninja and ccache not being found
which can throw an error if a binary is not found, stopping the build.
Use command -v as it is safer.

Also, switch from ?= to := as it was causing ninja not to be found
even with command -v. We'll print what binary the user is using so
that there is no confusion either.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>